requirements.txt
# QuantumNet Requirements - Python 3.12 Compatible
# Core Flask dependencies
Flask>=2.3.3
Flask-SocketIO>=5.3.6
Werkzeug>=2.3.7

# Database
SQLAlchemy>=2.0.21
Flask-SQLAlchemy>=3.0.5
psycopg2-binary>=2.9.9
redis>=5.0.1
eventlet>=0.36.1
Pillow>=10.3.0

# Cryptography
pycryptodome>=3.19.0
cryptography>=41.0.4
bcrypt>=4.1.2
pyotp>=2.9.0

# Machine Learning - Python 3.12 compatible versions
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.2

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Production
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Additional utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
